refactor(daco-26): rename app.tsx and change folder structure
- rename App.tsx to index.tsx
- move index.tsx to pages folder
- rename Home.tsx to HomePage.tsx
